requirements.txt
requests
pandas
fastapi
uvicorn
numpy
pyarrow
google-cloud-storage
python-multipart